+++ R/convert_ddi_gsed_12.R
@@ -1,12 +1,13 @@
-
-convert_ddi_gsed <- function(d, r, v) {
+convert_ddi_gsed_12 <- function(d, r, v) {
 
   # get BDS numbers and names of items
   bdsnum <- sort(unique(bdsreader::bds_gsed$bds))
   items <- bdsreader::bds_gsed$lex_gsed[bdsreader::bds_gsed$lex_gsed != ""]
 
   # premature return if there are no data
-  if (!(length(d$Contactmomenten) || length(d$ContactMomenten) || length(d$clientMeasurements))) {
+  if (!(length(d$Contactmomenten) ||
+        length(d$ContactMomenten) ||
+        length(d$clientMeasurements))) {
     return(data.frame(age = numeric(0)))
   }
   # premature return when no valid dates found
@@ -16,7 +17,9 @@ convert_ddi_gsed <- function(d, r, v) {
 
   # prepare the output matrices
   age <- as.numeric(round((r$dom - r$dob) / 365.25, 4L))
-  if (length(age) == 0L) {return(data.frame(age = numeric(0)))}
+  if (length(age) == 0L) {
+    return(data.frame(age = numeric(0)))
+  }
   w <- data.frame(
     age = age,
     matrix(NA, nrow = length(age), ncol = length(bdsnum) + length(items))
@@ -26,7 +29,7 @@ convert_ddi_gsed <- function(d, r, v) {
   # extract ddi data from bds-message
   # and convert to 0/1 scores
   for (i in bdsnum) {
-    if (v == 1 | v == 2) {
+    if (v == 1 || v == 2) {
       w[, as.character(i)] <- extract_field(d, i, v = v)
       next
     } else if (v == 3) {
@@ -35,7 +38,8 @@ convert_ddi_gsed <- function(d, r, v) {
         w[, as.character(i)] <- NA_real_
         next
       }
-      w[, as.character(i)] <- extr_bds[match(r$dom, ymd(extr_bds$date)), "value"]
+      w[, as.character(i)] <- extr_bds[match(r$dom, ymd(extr_bds$date)),
+                                       "value"]
     }
   }
 
@@ -43,9 +47,12 @@ convert_ddi_gsed <- function(d, r, v) {
     n <- which(bdsreader::bds_gsed$lex_gsed == item)
     type <- bdsreader::bds_gsed[n, "type"]
     bds <- bdsreader::bds_gsed[n, c("bds", "bdsr", "bdsl")]
-    w[, item] <- switch(type,
-      g1 = recode(w[, as.character(bds[1L])], `1` = 1L, `2` = 0L, .default = NA_integer_),
-      m1 = recode(w[, as.character(bds[1L])], `1` = 1L, `2` = 0L, `3` = 1L, .default = NA_integer_),
+    w[, item] <- switch(
+      type,
+      g1 = recode(w[, as.character(bds[1L])], `1` = 1L, `2` = 0L,
+                  .default = NA_integer_),
+      m1 = recode(w[, as.character(bds[1L])], `1` = 1L, `2` = 0L, `3` = 1L,
+                  .default = NA_integer_),
       g2 = {
         bdsr <- w[, as.character(bds[2L])]
         bdsl <- w[, as.character(bds[3L])]
@@ -64,18 +71,18 @@ convert_ddi_gsed <- function(d, r, v) {
     )
   }
 
-  # age-dependent post-processing to split items that map onto the same BDS number
-  # we must remove scores that map onto the same BDS number at multiple ages (962, 986)
-
-  # if age is missing, temporarily set it to -1 to evade error
+  # If age is missing, temporarily set it to -1 to evade error
   # "missing values are not allowed in subscripted assignments of data frames"
   age <- w$age
   age[is.na(age)] <- -1
 
+  # Age-dependent post-processing to split items that map onto the
+  # same BDS number. We must remove scores that map onto the same BDS
+  # number at multiple ages (962, 986).
   w[age < 0.5   / 365.25 | age > 42.5  / 365.25, "ddigmd155"] <- NA_real_
   w[age < 42.5  / 365.25 | age > 102.5 / 365.25, "ddigmd255"] <- NA_real_
   w[age < 102.5 / 365.25 | age > 146.5 / 365.25, "ddigmd355"] <- NA_real_
-  w[age < 146.5 / 365.25 | age > 1 , "ddigmd055"] <- NA_real_
+  w[age < 146.5 / 365.25 | age > 1, "ddigmd055"] <- NA_real_
 
   w[age < 0.75 | age >= 1.75, "ddigmd068"] <- NA_real_
   w[age < 1.75 | age >= 2.50, "ddigmd168"] <- NA_real_