+++ R/extract.R
@@ -28,23 +28,24 @@ extract_dob <- function(d, which = "00", v = 1) {
     return(return(as.Date(NA)))
   }
 
-  if (v == 3) p <- split(p, f = (1:nrow(p)))
+  if (v == 3) p <- split(p, f = seq_len(nrow(p)))
   p <- p[sapply(p, length) > 0]
-  for (i in 1L:length(p)) {
+  for (i in seq_along(p)) {
     pp <- p[[i]]
     parent <- switch(v,
                      pp[pp$Bdsnummer == 62L, "Waarde"],
                      pp[pp$ElementNummer == 62L, "Waarde"],
                      pp[pp$nestingBdsNumber == 62L, "nestingCode"]
-           )
+    )
     if (is.null(parent)) next
     if (parent == which) {
       # for v3 the parent codes are 'NFTH' (natural father) and "NMTH" (natural
       # mother)
-      dobp <- switch(v,
-                     ymd(pp[pp$Bdsnummer == 63L, 2L]),
-                     ymd(pp[pp$ElementNummer == 63L, 2L]),
-                     ymd(pp$clientDetails[[1]][pp$clientDetails[[1]]$bdsNumber == 63L, 2L])
+      dobp <- switch(
+        v,
+        ymd(pp[pp$Bdsnummer == 63L, 2L]),
+        ymd(pp[pp$ElementNummer == 63L, 2L]),
+        ymd(pp$clientDetails[[1]][pp$clientDetails[[1]]$bdsNumber == 63L, 2L])
       )
       if (!length(dobp)) {
         return(as.Date(NA))
@@ -100,11 +101,12 @@ extract_agep_v2 <- function(dob, dobp) {
 
 # For ContactMomenten
 extract_field <- function(d, f = 245L, v = 1) {
-  if (v == 1){
+  if (v == 1) {
     z <- d$Contactmomenten[[2L]]
     as.numeric(unlist(lapply(z, function(x, f2 = f) {
       ifelse("Waarde" %in% names(x), x[x$Bdsnummer == f2, "Waarde"], NA)
-    })))
+    }
+    )))
   } else if (v == 2) {
     z <- d$ContactMomenten[[2L]]
     as.numeric(unlist(lapply(z, function(x, f2 = f) {
@@ -129,6 +131,7 @@ extract_field2 <- function(d, f, v = 1) {
     b <- d[["ClientGegevens"]]
     if (!length(b)) return(NA_real_)
     v <- b[b$ElementNummer == f & !is.na(b$ElementNummer), "Waarde"]
+    v <- ifelse(v == "NA", NA_real_, v)
   } else if (v == 3) {
     b <- d[["clientDetails"]]
     if (!length(b)) return(NA_real_)
@@ -147,8 +150,8 @@ extract_field3 <- function(d, f, which_parent = "02", v = 1) {
   if (length(p) == 0) {
     return(NA)
   }
-  if (v == 3) p <- split(p, f = (1:nrow(p)))
-  for (i in 1L:length(p)) {
+  if (v == 3) p <- split(p, f = seq_len(nrow(p)))
+  for (i in seq_along(p)) {
     pp <- p[[i]]
     parent <- switch(v,
                      pp[pp$Bdsnummer == 62L, "Waarde"],
@@ -156,10 +159,11 @@ extract_field3 <- function(d, f, which_parent = "02", v = 1) {
                      pp[pp$nestingBdsNumber == 62L, "nestingCode"])
     if (is.null(parent)) next
     if (parent == which_parent) {
-      wp <- switch(v,
-                   pp[pp$Bdsnummer == f, 2L],
-                   pp[pp$ElementNummer == f, 2L],
-                   pp$clientDetails[[1]][pp$clientDetails[[1]]$bdsNumber == f, 2L])
+      wp <- switch(
+        v,
+        pp[pp$Bdsnummer == f, 2L],
+        pp[pp$ElementNummer == f, 2L],
+        pp$clientDetails[[1]][pp$clientDetails[[1]]$bdsNumber == f, 2L])
       return(wp)
     }
   }