+++ R/internal.R
@@ -1,1 +1,0 @@
-is.url <- function(x) grepl("^https?://", x, useBytes = TRUE)