+++ R/throw_messages.R
@@ -4,12 +4,12 @@ throw_messages <- function(x) {
   }
 
   if (is.vector(x)) {
-    for (i in 1:length(x)) {
+    for (i in seq_along(x)) {
       message(x[i])
     }
   }
   if (is.data.frame(x)) {
-    for (i in 1:nrow(x)) {
+    for (i in seq_len(nrow(x))) {
       st <- paste0(
         "BDS ", format(x[i, 1L], width = 3, justify = "right"), " (",
         x[i, 2L], "): Supplied: ",